+++ src/gems/expression/uses_sum_connections_on.py
